refactor(board): rename hooks to new mutation-query notation
