+++ src/modules/core.fetching-hooks/boards/index.ts
@@ -1,5 +1,5 @@
-export * from "./use-delete-board-by-id";
-export * from "./use-get-all-boards";
-export * from "./use-get-board-by-id";
-export * from "./use-post-create-board";
-export * from "./use-put-update-board-by-id";
+export * from "./use-mutation-create-board";
+export * from "./use-mutation-delete-board-by-id";
+export * from "./use-mutation-update-board-by-id";
+export * from "./use-query-all-boards";
+export * from "./use-query-board-by-id";
