+++ src/modules/core.fetching-hooks/boards/use-mutation-delete-board-by-id.ts
@@ -2,19 +2,24 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { MutationKey, QueryKey } from "@/modules/fetcher";
 import {
   useRepositories,
-  type IMutationCreateBoardIn,
+  type IMutationDeleteBoardByIdIn,
 } from "@/modules/repositories/app";
 
-export function usePostCreateBoard() {
+export function useMutationDeleteBoardById() {
   const queryClient = useQueryClient();
   const { board } = useRepositories();
 
   return useMutation({
-    mutationKey: [MutationKey.POST_CREATE_BOARD],
-    mutationFn: (args: IMutationCreateBoardIn) => board.createBoard(args),
-    onSettled: () => {
+    mutationKey: [MutationKey.DELETE_BOARD],
+    mutationFn: (args: IMutationDeleteBoardByIdIn) =>
+      board.deleteBoardById(args),
+    onSettled: (_, __, { id }) => {
       void queryClient.invalidateQueries({
         queryKey: [QueryKey.GET_ALL_BOARDS],
+      });
+
+      void queryClient.invalidateQueries({
+        queryKey: [QueryKey.GET_BOARD_BY_ID, id],
       });
     },
   });
